Splitting code into multiple files for clarity
+ Better code documentation
